improvements
Fixed bugs

Updated API paths
